Refactor moving files closer to their usage
