+++ frontend/src/metabase/common/hooks/index.ts
@@ -9,7 +9,6 @@ export * from "./use-database-list-query";
 export * from "./use-database-query";
 export * from "./use-metric-list-query";
 export * from "./use-metric-query";
-export * from "./use-most-recently-viewed-dashboard";
 export * from "./use-popular-item-list-query";
 export * from "./use-question-list-query";
 export * from "./use-question-query";
