--- frontend/src/metabase/common/hooks/use-most-recently-viewed-dashboard/index.ts
@@ -1,1 +1,0 @@
-export * from "./use-most-recently-viewed-dashboard";
